% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_areasqkm}
\alias{add_areasqkm}
\title{Compute area in square kilometers}
\usage{
add_areasqkm(x)
}
\arguments{
\item{x}{POLYGON sf object}
}
\value{
numeric vector
}
\description{
Compute area in square kilometers
}
