% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{add_length_map}
\alias{add_length_map}
\title{Add Length Mapping from VAA}
\usage{
add_length_map(flowpaths, length_table)
}
\arguments{
\item{flowpaths}{an sf object}
}
\value{
data.frame
}
\description{
Add Length Mapping from VAA
}
