% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_utils.R
\name{add_lengthkm}
\alias{add_lengthkm}
\title{Compute km length
Short hand for safely computing length in km and returning as numeric vector.}
\usage{
add_lengthkm(x)
}
\arguments{
\item{x}{sf object}
}
\value{
numeric vector
}
\description{
Compute km length
Short hand for safely computing length in km and returning as numeric vector.
}
