% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_measures}
\alias{add_measures}
\title{Add Measures to Flowlines and Catchments}
\usage{
add_measures(flowpaths, cat)
}
\arguments{
\item{flowpaths}{sf object (LINESTRING)}

\item{cat}{sf object (POLYGON)}
}
\value{
list
}
\description{
Add Measures to Flowlines and Catchments
}
