% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_thresholds.R
\name{aggregate_by_thresholds}
\alias{aggregate_by_thresholds}
\title{Aggregate Refactored Hydrofabric by Thresholds}
\usage{
aggregate_by_thresholds(
  fl = NULL,
  catchments = NULL,
  gpkg = NULL,
  fl_name = "refactored_flowpaths",
  cat_name = "refactored_catchments",
  ideal_size = 10,
  min_area_sqkm = 3,
  min_length_km = 1,
  condition = "or"
)
}
\arguments{
\item{fl}{a flowline sf object}

\item{catchments}{a catchment sf object}

\item{gpkg}{A path to a geopackage with refactored output}

\item{fl_name}{The name of the refactored flowline layer in \code{gpkg}}

\item{cat_name}{The name of the refactored catchment layer in \code{gpkg}}

\item{ideal_size}{The ideal size of output hydrofabric catchments}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}

\item{condition}{How should headwaters be collapsed? Those where the area AND length are less then
the prescribed thresholds, or where the area OR length are less then the thresholds
(options are "and" or "or").}
}
\value{
list object
}
\description{
A complete network aggregating workflow has been packaged
into this function. Builds a set of normalized catchment-flowpaths from
input flowline and catchment layer that meet specified area and length thresholds.
}
\details{
This is a convenient wrapper function that implements four phases
of the network aggregation workflow: merge along mainstems, correcting parallel flow, collapsing headwaters.
}
\seealso{
The following three functions are used in the \code{aggregate_by_thresholds} workflow.
\enumerate{
\item \code{\link{merge_along_mainstem}}
\item \code{\link{collapse_headwaters}}
}
}
