% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfe_noahowp_summary.R
\name{aggregate_cfe_noahowp}
\alias{aggregate_cfe_noahowp}
\title{Aggregated CFE variables frome NWM}
\usage{
aggregate_cfe_noahowp(
  gpkg = NULL,
  dir = NULL,
  catchment_name = "aggregate_divides",
  flowline_name = "aggregate_flowpaths",
  precision = 9,
  add_to_gpkg = TRUE
)
}
\arguments{
\item{gpkg}{a geopackage with aggregation units}

\item{dir}{directory of expected nwm_grid_file (see hyAggregate::get_nwm_grids())}

\item{catchment_name}{the layer name of the aggregation units}

\item{flowline_name}{the layer name of the flowpath units}

\item{precision}{the precision of the computations}

\item{single_layer}{should only the top layer of a multilayer parameter be processed?}
}
\description{
Aggregated CFE variables frome NWM
}
