% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_utils.R
\name{aggregate_network_by_index}
\alias{aggregate_network_by_index}
\title{Aggregate Network by Index}
\usage{
aggregate_network_by_index(fl, cat, index_table, grouping_id = "set")
}
\arguments{
\item{fl}{a LINESTRING \code{sf} flowlines object}

\item{cat}{a POLYGON \code{sf} catchments object}

\item{index_table}{a table of network metadata including
c(grouping_id, 'ID',  'toID', 'Hydroseq', 'member_COMID', 'LevelPathID')}

\item{grouping_id}{the name of the index table containing the grouping variable}
}
\value{
a list containing aggregated and validated flowline and catchment \code{sf} objects
}
\description{
When provided with an index table, return a set of aggregated
and  normalized catchment-flowpaths
}
