% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_network_to_distribution}
\alias{aggregate_network_to_distribution}
\title{Aggregate Network to Distribution}
\usage{
aggregate_network_to_distribution(
  gpkg = NULL,
  catchment_name = "refactored_divides",
  flowpath_name = "refactored_flowpaths",
  ideal_size_sqkm = 10,
  min_length_km = 1,
  min_area_sqkm = 3,
  nexus_topology = TRUE,
  routelink_path = NULL,
  nexus_locations = NULL,
  outfile = NULL,
  verbose = TRUE,
  overwrite = FALSE,
  cache_file = NULL,
  log_file = NULL
)
}
\arguments{
\item{ideal_size_sqkm}{The ideal size of catchments (default = 10 sqkm)}

\item{min_length_km}{The minimum allowable length of flowpath features (default = 1 km)}

\item{min_area_sqkm}{The minimum allowable area of catchment features (default = 3 sqkm)}

\item{nexus_topology}{should a hy-features cat-->nex topology be enforced? default = TRUE}

\item{nexus_locations}{a data.frame with columns specifiying the ID, and the nexus type.}

\item{outfile}{of not NULL, where to write the output files}

\item{verbose}{print status updates. Default = TRUE}

\item{overwrite}{overwrite existing gf file. Default is FALSE}

\item{gf}{A path to a hydrofabric file (see US reference fabric)}
}
\value{
if outfile = TRUE, a file path, else a list object
}
\description{
This function aggregated a network to a desired size distribution while
enforcing minimum flowpath legnths and catchment area. Addtionally a set of explicit nexus
locations can be provided over which the netwrok cannot be aggregated.
}
