% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_sets}
\alias{aggregate_sets}
\title{Aggregate Sets by Index Table}
\usage{
aggregate_sets(flowpaths, cat, index_table)
}
\arguments{
\item{flowpaths}{LINESTRING flowpaths}

\item{cat}{POLYGON catchments}

\item{index_table}{index table to aggregate with}
}
\value{
a list of catchments and flowpaths that have been validated
}
\description{
Aggregate Sets by Index Table
}
