% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_id}
\alias{assign_id}
\title{Index a Vector by cumulative Sum}
\usage{
assign_id(a, athres)
}
\arguments{
\item{a}{a vector of values}

\item{athres}{the ideal size of each index. Cummulative sums will get as close to this value without exceeding it}
}
\value{
a vector of length(a)
}
\description{
Index a Vector by cumulative Sum
}
