% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_by_thresholds.R
\name{build_network_list}
\alias{build_network_list}
\title{Build a network list object}
\usage{
build_network_list(
  gpkg = NULL,
  fl_name = "refactored_flowpaths",
  cat_name = "refactored_catchments"
)
}
\arguments{
\item{gpkg}{A path to a geopackage with refactored output}

\item{fl_name}{The name of the refactored flowline layer in \code{gpkg}}

\item{cat_name}{The name of the refactored catchment layer in \code{gpkg}}
}
\value{
a list containing flowline and catchment \code{sf} objects
}
\description{
A network_list is a list of catchment and flowpaths \code{sf} objects that are validated.
Validations include that all required names are present; that all toIDs are actual
present IDs and that the flowline network is a DAG (Directed acyclic graph).
}
