% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_network_validity}
\alias{check_network_validity}
\title{Check Network Validity
\strong{INTERNAL} function that validates a flowpath and catchment network}
\usage{
check_network_validity(network_list, term_cut = 1e+09, check = TRUE)
}
\arguments{
\item{term_cut}{cutoff integer to define terminal IDs}
}
\value{
a list containing flowline and catchment \code{sf} objects
}
\description{
Check Network Validity
\strong{INTERNAL} function that validates a flowpath and catchment network
}
