% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_headwaters.R
\name{collapse_headwaters}
\alias{collapse_headwaters}
\title{Collapse headwaters}
\usage{
collapse_headwaters(
  network_list,
  min_area_sqkm = 3,
  min_length_km = 1,
  condition = "or"
)
}
\arguments{
\item{network_list}{a list containing flowline and catchment \code{sf} objects}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}

\item{condition}{How should headwaters be collapsed? Those where the area AND length are less then
the prescribed thresholds, or where the area OR length are less then the thresholds
(options are "and" or "or").}
}
\value{
list of sf objects
}
\description{
Headwaters are those segments in which there are no inflows (!ID \%in\% toID).
This function removes and normalizes the network list to collapse headwater locations that violate
prescribed minimum length and area thresholds.
}
