% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{collapse_hw}
\alias{collapse_hw}
\title{Collapse Headwaters}
\usage{
collapse_hw(network_list, min_area_sqkm = 3, min_length_km = 1, verbose = TRUE)
}
\arguments{
\item{network_list}{a list containing flowpath and catchment \code{sf} objects}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}

\item{verbose}{should messages be emitted?}
}
\value{
a list containing flowpath and catchment \code{sf} objects
}
\description{
This function identifies small (pathlength or area) headwater catchments and
collapses them into the existing network uptil none remain.
Headwaters are those segments in which there are no inflows (!ID \%in\% toID).
}
