% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{collapse_network_inward}
\alias{collapse_network_inward}
\title{Collapse Network}
\usage{
collapse_network_inward(
  network_list,
  min_area_sqkm,
  min_length_km,
  verbose = TRUE
)
}
\arguments{
\item{network_list}{a list containing flowline and catchment \code{sf} objects}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}
}
\value{
list of sf objects
}
\description{
Headwaters are those segments in which there are no inflows (!ID \%in\% toID).
This function removes and normalizes the network list to collapse headwater locations that violate
prescribed minimum length and area thresholds.
}
