% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfe_noahowp_summary.R
\name{correct_nwm_spatial}
\alias{correct_nwm_spatial}
\title{Correct Degenerate NWM files}
\usage{
correct_nwm_spatial(path, ...)
}
\arguments{
\item{path}{path to NWM file}

\item{subds}{subdatasets to extract}
}
\value{
SpatRast object
}
\description{
Degenerate = having lost the physical, mental, or moral qualities considered normal and desirable; showing evidence of decline. NWM files do not store extents or projections...
}
