% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_utils.R
\name{cs_group}
\alias{cs_group}
\title{Cumulative sum area grouping}
\usage{
cs_group(a, athres)
}
\arguments{
\item{a}{a vector of areas}

\item{athres}{a threshold, or target, cumulative size}
}
\value{
a vector of length(a) containing grouping indexes
}
\description{
This function takes a vector of area's and returns a
grouping vector that aims to combine then towards an ideal aggregate size (threshold)
}
