% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtools_download_upload.R
\name{.getExtension}
\alias{.getExtension}
\title{Extract File Extension}
\usage{
.getExtension(x, prefix = "")
}
\arguments{
\item{x}{a file path}

\item{prefix}{character string to precede extracted extension.  default = "". (Usfull if you want to keep the ".")}
}
\value{
character string
}
\description{
Extract File Extension
}
\details{
returns file extension
}
