% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_utils.R
\name{drop_artifical_splits}
\alias{drop_artifical_splits}
\title{Drop Artifical Splits}
\usage{
drop_artifical_splits(fl)
}
\arguments{
\item{fl}{a LINESTRING \code{sf} flowlines object}
}
\value{
a LINESTRING \code{sf} flowlines object
}
\description{
Removes Orphan flowpaths that do not connect to the network
and that are represented as a MULTILINESTRING object
}
