% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_headwaters.R
\name{find_and_remove_detached}
\alias{find_and_remove_detached}
\title{On a rare occasion, the combination of HY_Refactor water
flying can result in a flowpath that (A) does not have any inflows, (B) violates lengths/area assumptions,
(C) is NOT a headwater. In these cases a segement within the network can get removed leaving isolated/detached flowpaths.
This functions is a first, but probably not ideal, stab at cleaning these up.}
\usage{
find_and_remove_detached(fl, cats, term_cut)
}
\arguments{
\item{fl}{a flowline sf object}

\item{cats}{a catchment sf object}

\item{term_cut}{cutoff integer to define terminal IDs}
}
\value{
a validated network list
}
\description{
On a rare occasion, the combination of HY_Refactor water
flying can result in a flowpath that (A) does not have any inflows, (B) violates lengths/area assumptions,
(C) is NOT a headwater. In these cases a segement within the network can get removed leaving isolated/detached flowpaths.
This functions is a first, but probably not ideal, stab at cleaning these up.
}
