% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtools_download_upload.R
\name{find_pu}
\alias{find_pu}
\title{Find Processing Unit (Vector or Raster)}
\usage{
find_pu(location, pu = "vpu")
}
\arguments{
\item{location}{sf object}

\item{pu}{either "vpu" or "rpu" (default is "vpu")}
}
\value{
intersection processing units (sf object)
}
\description{
Find Processing Unit (Vector or Raster)
}
