% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_lookup_table}
\alias{generate_lookup_table}
\title{Generate Lookup table for Aggregated Nextwork}
\usage{
generate_lookup_table(
  gpkg = NULL,
  flowpath_name = "aggregate_flowpaths",
  refactored_fabric = NULL
)
}
\arguments{
\item{gpkg}{to the aggregated network. The look up table will be added here as a layer called "lookup_table"}

\item{flowpath_name}{the name of the flowpath layer in gpkg}

\item{refactored_fabric}{the gpkg for the corresponding VPU reference fabric}
}
\value{
data.frame
}
\description{
Generate Lookup table for Aggregated Nextwork
}
