% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geos_union_polygon_hyaggregate}
\alias{geos_union_polygon_hyaggregate}
\title{Fast POLYGON Union}
\usage{
geos_union_polygon_hyaggregate(poly, ID)
}
\arguments{
\item{poly}{sf POLYGON object}

\item{ID}{the column name over which to union geometries}
}
\value{
sf object
}
\description{
This is significantly faster then sf::st_union or summarize
}
