% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{get_catchment_data}
\alias{get_catchment_data}
\title{Get Catchment Data}
\usage{
get_catchment_data(catchment, catchment_edge_list, catchment_prefix = "cat-")
}
\arguments{
\item{catchment}{sf data.frame containing hyRefactor or hyAggregate output.}

\item{wb_prefix}{character prefix for waterbody IDs}

\item{terminal_wb_prefix}{character prefix for terminal waterbody IDs}

\item{cutoff}{terminal IDs begin above a defined threshold}
}
\value{
data.frame
}
\description{
get a edge list for waterbodies
}
