% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{get_catchment_edges_terms}
\alias{get_catchment_edges_terms}
\title{Get catchment edge list}
\usage{
get_catchment_edges_terms(
  flowpaths,
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnex-",
  catchment_prefix = "cat-",
  cutoff = 100000000
)
}
\arguments{
\item{flowpaths}{sf data.frame containing hyRefactor or hyAggregate output.}

\item{nexus_prefix}{character prefix for nexus IDs}

\item{terminal_nexus_prefix}{character prefix for terminal nexus IDs}

\item{catchment_prefix}{character prefix for catchment IDs}

\item{cutoff}{terminal IDs begin above a defiend threshold}
}
\value{
data.frame
}
\description{
get a edge list for catchments
}
