% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{get_nexus}
\alias{get_nexus}
\title{get nexuses}
\usage{
get_nexus(
  fline,
  term_cut = 1e+09,
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-"
)
}
\arguments{
\item{fline}{sf data.frame NHDPlus Flowlines or hyRefactor output.}

\item{nexus_prefix}{character prefix for nexus IDs}
}
\description{
get nexuses

get nexuses
}
