% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nexus_locations.R
\name{get_nexus_locations}
\alias{get_nexus_locations}
\title{Get Nexus Locations}
\usage{
get_nexus_locations(fp, term_cut = 100000000)
}
\arguments{
\item{fp}{a sf flowpath object}

\item{term_cut}{cutoff integer to define terminal IDs}
}
\value{
POINT sf object
}
\description{
Get Nexus Locations
}
