% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwm_grids.R
\name{get_nwm_grids}
\alias{get_nwm_grids}
\title{Get NWM hydrologcic grids}
\usage{
get_nwm_grids(dir = NULL, spatial = FALSE, download_cache = TRUE)
}
\arguments{
\item{dir}{directory where file should exist}

\item{spatial}{should the returned object be a file path (spatial= FALSE, default) or SpatRast (spatial = TRUE)}

\item{download_cache}{if the file does not exist locally, should it be downloaded (TRUE) or built from raw data (FALSE)}
}
\value{
a file path (spatial = FALSE) or SpatRast object (spatial = TRUE)
}
\description{
Base Nextgen requires summary data from 17 layers across the soilproperties_CONUS_FullRouting
GWBUCKPARM_CONUS_FullRouting, and wrfinput_CONUS NWM NetCDF files. These three files have been combined and the needed layers extracted.
If this dataset doesn't exist locally, it can be built using this function (download_cache = FALSE) or downloaded from ScienceBase (download_cache = FALSE).
The spatial metadata in these files is incomplete and needs to be corrected. This can be done by passing the path returned by this function when spatial = FALSE to
\code{correct_nwm_spatial} or, spatial can be set to TRUE. Using data_cache = FALSE requires NCO to be installed!
}
