% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtools_download_upload.R
\name{get_reference_fabric}
\alias{get_reference_fabric}
\title{Download Reference Fabric Data by VPU ID}
\usage{
get_reference_fabric(
  VPU = "01",
  type = "refactored",
  dir = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{VPU}{a VPU ID}

\item{type}{either 'refactored' (default) or 'reference'}

\item{dir}{directory path to save data to}

\item{overwrite}{should existing files be overwritten? (default = FALSE)}
}
\value{
file path
}
\description{
Download Reference Fabric Data by VPU ID
}
