% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{get_routelink}
\alias{get_routelink}
\title{Routlink Attribute Subset}
\usage{
get_routelink(atts = NULL, path = get_routelink_path(), build = TRUE)
}
\arguments{
\item{atts}{character The variable names you would like, always includes comid}

\item{path}{character path where the file should be saved. Default is a persistent system data as retrieved by hyAggregate_data_dir. Also see: get_routlink_path}

\item{build}{if TRUE, and the file does not exist, should it be built?}
}
\value{
character
}
\description{
Routlink Attribute Subset
}
