% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{get_routelink_path}
\alias{get_routelink_path}
\title{Get Routelink Path}
\usage{
get_routelink_path(dir = hyAggregate_data_dir(), build = TRUE)
}
\arguments{
\item{dir}{if not supplies will default to \code{get("ngen_dat_dir", envir = hyAggregate_env)}}

\item{build}{if TRUE, and the file does not exist, should it be built?}
}
\value{
character
}
\description{
Get Routelink Path
}
