% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_UT_geoms.R
\name{get_UT_geoms}
\alias{get_UT_geoms}
\title{Extract upstream network from reference fabric}
\usage{
get_UT_geoms(network, reference_fabric_dir, comid)
}
\arguments{
\item{network}{enhd_nhdplusatts data.frame}

\item{reference_fabric_dir}{path to refernece network directory}

\item{comid}{Origin of network trace (COMID)}
}
\value{
list containing catchment and flowpath networks
}
\description{
Extract upstream network from reference fabric
}
