% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyAggregate.R
\docType{package}
\name{hyAggregate-package}
\alias{hyAggregate}
\alias{hyAggregate-package}
\title{hyAggregate: Hydrologic Network Aggregating Tools Based on HY_Features}
\description{
A collection of tools for normalizing and aggregating the size of catchments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mikejohnson51/hyAggregate}
  \item Report bugs at \url{https://github.com/mikejohnson51/hyAggregate/issues}
}

}
\author{
\strong{Maintainer}: Mike Johnson \email{mikecp11@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
