% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_subsetting.R
\name{layer_exists}
\alias{layer_exists}
\title{Check if geopackage layer exists}
\usage{
layer_exists(gpkg, name)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{name}{layer name}
}
\value{
boolean
}
\description{
Check if geopackage layer exists
}
