% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/routing.R
\name{length_average_routelink}
\alias{length_average_routelink}
\title{Length Average Routelink Variables}
\usage{
length_average_routelink(
  flowpaths,
  rl_vars = c("comid", "Qi", "MusK", "MusX", "n", "So", "ChSlp", "BtmWdth", "time",
    "Kchan", "nCC", "TopWdthCC", "TopWdth"),
  rl_path = get_routelink_path()
)
}
\arguments{
\item{flowpaths}{sf LINESTRING}

\item{rl_vars}{routelink variables}

\item{rl_path}{routelink path (see get_routelink_path())}
}
\value{
data.frame
}
\description{
Length Average Routelink Variables
}
