% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor_wrapper.R
\name{length_average_routlink}
\alias{length_average_routlink}
\title{Assign Length Weighted Flowpath Attributes from Routelink file}
\usage{
length_average_routlink(flowpaths, rl_vars, rl_path)
}
\arguments{
\item{flowpaths}{flowpaths with a\code{member_COMID} attribute}

\item{rl_vars}{RouteLink variable to append}

\item{rl_path}{Path to RouteLink}
}
\value{
flowpaths with length weighted RouteLink Attributes
}
\description{
This uses the hyRefactor::add_lengthmap to determine the weighted
length of comtributing NHD COMIDs to any refactored/aggregated flowpath.
These are then used to assign length averaged values from NWM RouteLink file.
}
