% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_along_mainstem.R
\name{merge_along_mainstem}
\alias{merge_along_mainstem}
\title{Aggregate along network mainstems}
\usage{
merge_along_mainstem(
  network_list,
  ideal_size = 10,
  min_area_sqkm = 3,
  min_length_km = 1
)
}
\arguments{
\item{network_list}{a list containing flowline and catchment \code{sf} objects}

\item{ideal_size}{The ideal size of output hydrofabric catchments}

\item{min_area_sqkm}{The minimum allowable size of the output hydrofabric catchments}

\item{min_length_km}{The minimum allowable length of the output hydrofabric flowlines}
}
\value{
a list containing aggregated and validated flowline and catchment \code{sf} objects
}
\description{
Given a set of ideal catchment sizes, plus the
minimum allowable catchment size and segment length, aggregate the network along mainstems.
}
