% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nexus_from_poi}
\alias{nexus_from_poi}
\title{Extract nexus locations for Reference POIs}
\usage{
nexus_from_poi(
  gpkg,
  type = c("HUC12", "Gages", "TE", "NID", "WBIn", "WBOut"),
  verbose = TRUE
)
}
\arguments{
\item{gpkg}{a reference hydrofabric gpkg}

\item{type}{the type of desired POIs}

\item{verbose}{should messages be emitted?}
}
\value{
data.frame with ID, type columns
}
\description{
Extract nexus locations for Reference POIs
}
