% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_hydrofabric_package}
\alias{read_hydrofabric_package}
\title{Read Catchments and Flowpaths from Geopackage
Convienence function for reading two layers into a list}
\usage{
read_hydrofabric_package(gpkg, catchment_name, flowpath_name, crs = NULL)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{catchment_name}{name of catchment layer}

\item{flowpath_name}{name of flowpath layer}

\item{crs}{desired CRS, if NULL they stay as read}
}
\value{
list
}
\description{
Read Catchments and Flowpaths from Geopackage
Convienence function for reading two layers into a list
}
