% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_topology.R
\name{realign_topology}
\alias{realign_topology}
\title{Realign topology to Hy Feature standard}
\usage{
realign_topology(network_list, term_cut = 100000000)
}
\arguments{
\item{network_list}{a list containing flowline and catchment \code{sf} objects}

\item{term_cut}{cutoff integer to define terminal IDs}
}
\value{
a list containing aggregated and validated flowline and catchment \code{sf} objects
}
\description{
Realign topology to Hy Feature standard
}
