% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbtools_download_upload.R
\name{upload_aggregate_gpkg}
\alias{upload_aggregate_gpkg}
\title{Authenticated Upload to ScienceBase}
\usage{
upload_aggregate_gpkg(x, item = "629a4246d34ec53d276f446d", username, password)
}
\arguments{
\item{x}{A string vector of paths to files to be uploaded}

\item{item}{a character ScienceBase ID corresponding to the item (default = '629a4246d34ec53d276f446d')}

\item{username}{Sciencebase username, prompts user if not supplied}

\item{password}{Sciencebase password, prompts user if not supplied}
}
\description{
Authenticated Upload to ScienceBase
}
