% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_download_upload.R
\name{upload_to_aws}
\alias{upload_to_aws}
\title{Authenticated Upload to s3
You must authenticate s3 for this working session to use this function. Also,
the aws.s3 package is required on your machine and is NOT a hyAggregate dependency.}
\usage{
upload_to_aws(
  path,
  bucket = "formulations-dev",
  prefix = "hf_1.0",
  verbose = TRUE
)
}
\arguments{
\item{path}{a path to a file or directory}

\item{bucket}{the s3 bucket to upload to}

\item{prefix}{an optional file object prefix (think of sub directory)}

\item{verbose}{should messages be emitted?}
}
\description{
Authenticated Upload to s3
You must authenticate s3 for this working session to use this function. Also,
the aws.s3 package is required on your machine and is NOT a hyAggregate dependency.
}
