% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngen_export.R
\name{write_ngen_dir}
\alias{write_ngen_dir}
\title{Write NextGen files from GPKG}
\usage{
write_ngen_dir(
  gpkg,
  dir = NULL,
  catchment_name = "aggregate_divides",
  flowpath_name = "aggregate_flowpaths",
  export_shapefiles = FALSE,
  verbose = TRUE,
  overwrite = FALSE
)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{dir}{directory to write data to, if NULL, then its written at the same level as the gpkg}

\item{catchment_name}{name of catchment layer}

\item{flowpath_name}{name of flowpath layer}
}
\description{
Write NextGen files from GPKG
}
