% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{aggregate_network_to_distribution}
\alias{aggregate_network_to_distribution}
\title{Aggregate Network}
\usage{
aggregate_network_to_distribution(
  gf = NULL,
  flowpaths = NULL,
  catchments = NULL,
  ideal_size_sqkm = 10,
  min_length_km = 1,
  min_area_sqkm = 3,
  outfile = NULL,
  verbose = TRUE,
  overwrite = FALSE,
  nexus_topology = TRUE,
  routelink_path = NULL
)
}
\arguments{
\item{gf}{A path to a refactored geofarbric file (see US reference fabric)}

\item{flowpaths}{an sf object}

\item{catchments}{an sf object}

\item{ideal_size_sqkm}{The ideal size of catchments (default = 10 sqkm)}

\item{min_length_km}{The minimum allowable length of flowpaths (default = 1 km)}

\item{min_area_sqkm}{The minimum allowable size of catchments (default = 3 sqkm)}

\item{outfile}{of not NULL, where to write the output files}

\item{verbose}{print status updates. Default = TRUE}

\item{overwrite}{overwrite existing gf file. Default is FALSE}

\item{nexus_topology}{should a hy-features cat-->nex topology be enforced? default = TRUE}
}
\value{
if outfile = TRUE, a file path, else a list object
}
\description{
Aggregate Network
}
