% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{get_nexus}
\alias{get_nexus}
\title{get nexuses}
\usage{
get_nexus(
  fp,
  term_cut = 1000000000,
  nexus_prefix = "nex-",
  terminal_nexus_prefix = "tnx-"
)
}
\arguments{
\item{nexus_prefix}{character prefix for nexus IDs}

\item{terminal_nexus_prefix}{character prefix for ternimal nexus IDs}

\item{fline}{sf data.frame NHDPlus Flowlines or hyRefactor output.}
}
\description{
get nexuses
}
