% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_feature_functions.R
\name{get_waterbody_edges_terms}
\alias{get_waterbody_edges_terms}
\title{Get waterbody edge list}
\usage{
get_waterbody_edges_terms(
  flowpaths,
  wb_prefix = "wb-",
  terminal_wb_prefix = "twb-",
  cutoff = 100000000
)
}
\arguments{
\item{flowpaths}{sf data.frame containing hyRefactor or hyAggregate output.}

\item{wb_prefix}{character prefix for waterbody IDs}

\item{terminal_wb_prefix}{character prefix for terminal waterbody IDs}

\item{cutoff}{terminal IDs begin above a defined threshold}
}
\value{
data.frame
}
\description{
get a edge list for waterbodies
}
