% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_subsetting.R, R/utils.R
\name{layer_exists}
\alias{layer_exists}
\title{Check if geopackage layer exists}
\usage{
layer_exists(gpkg, name)

layer_exists(gpkg, name)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{name}{name of layer to check}
}
\value{
boolean

logical
}
\description{
Check if geopackage layer exists

Check if a geopackage and layer exists
This function checks if a layer exists in a geopackage
}
