% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{network_is_dag}
\alias{network_is_dag}
\title{Check if network is DAG
Checks in a \code{sf} flowline network is a DAG (Directed acyclic graph).}
\usage{
network_is_dag(fl, ID = "id", toID = "toid")
}
\arguments{
\item{fl}{a LINESTRING \code{sf} flowlines object}

\item{ID}{the name of the ID column in \code{fl}}

\item{toID}{the name of the toID column in \code{fl}}
}
\value{
boolean
}
\description{
Check if network is DAG
Checks in a \code{sf} flowline network is a DAG (Directed acyclic graph).
}
