% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_subsetting.R
\name{subset_network}
\alias{subset_network}
\title{Subset the upstream protion of a network}
\usage{
subset_network(gpkg, origin, mainstem = FALSE)
}
\arguments{
\item{gpkg}{path to a hydrofabric}

\item{origin}{the ID to begin navigatipn}
}
\value{
mainstem should only the mainstem flowpath be retruned (default = FALSE)
}
\description{
Subset the upstream protion of a network
}
