% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngen_export.R
\name{write_ngen_dir}
\alias{write_ngen_dir}
\title{Write NextGen files from GPKG}
\usage{
write_ngen_dir(
  gpkg,
  catchment_name = "aggregate_catchments",
  flowpath_name = "aggregate_flowpaths",
  export_shapefiles = FALSE
)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{catchment_name}{name of catchment layer}

\item{flowpath_name}{name of flowpath layer}
}
\description{
Write NextGen files from GPKG
}
