% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngen_export.R
\name{write_shapefile_dir}
\alias{write_shapefile_dir}
\title{Write NextGen geopackage as directory of shapefiles}
\usage{
write_shapefile_dir(gpkg, dir)
}
\arguments{
\item{gpkg}{path to geopackage}

\item{dir}{directory path to create a 'shp' folder for output}
}
\description{
Write NextGen geopackage as directory of shapefiles
}
